requirement.txt
asgiref==3.4.1
awesome-slugify==1.6.5
beautifulsoup4==4.10.0
Django==4.0.1
django-bootstrap4==21.2
django-filter==21.1
django-search-views==0.3.7
Pillow==9.0.0
regex==2022.1.18
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
Unidecode==0.4.21
